requirements.txt
Flask==3.1.2
mysql-connector-python==9.0.0
gunicorn==23.0.0
